+++ actix-web-lab/src/guard.rs
@@ -1,6 +1,3 @@
 //! Experimental route guards.
 //!
 //! Analogous to the `guard` module in Actix Web.
-
-#[allow(deprecated)]
-pub use crate::acceptable::Acceptable;